+++ ProceduralForest/ProceduralForest/src/Forest.h
@@ -29,7 +29,7 @@ public:
 	std::vector<vec3>* leafMatrixCol4;
 
 private:
-	void AddTree(glm::vec3 pos, float height, float maxDepth, float maxBranches);
+	void AddTree(glm::vec3 pos, float height, int maxDepth, int maxBranches);
 	void MakeBranches(const int maxDepth, int currentDepth, float currentHeight, int branches, float totalScale);
 	void CreateCylinder(int aSlices, float height, float topwidth, float bottomwidth);
 	void GenerateTerrain();
